render_matcher/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "render-matcher"
version = "0.1.0"
description = "CNN-based render vs reality image matching for 3D print verification"
readme = "README.md"
license = {text = "AGPL-3.0-or-later"}
requires-python = ">=3.9"
authors = [
    {name = "PrusaSlicer Contributors"}
]
keywords = ["cnn", "siamese-network", "image-matching", "3d-printing", "quality-control"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: GNU Affero General Public License v3 or later (AGPLv3+)",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Image Recognition",
]
dependencies = [
    "torch>=2.0",
    "torchvision>=0.15",
    "albumentations>=1.3",
    "tensorboard>=2.12",
    "pyyaml>=6.0",
    "opencv-python>=4.7",
    "tqdm>=4.65",
    "scikit-learn>=1.2",
    "numpy>=1.24",
]

[project.optional-dependencies]
dev = ["pytest", "black", "isort", "mypy"]

[project.urls]
Homepage = "https://github.com/prusa3d/PrusaSlicer"
Repository = "https://github.com/prusa3d/PrusaSlicer"

[project.scripts]
render-matcher-train = "render_matcher.trainer:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["render_matcher*"]

[tool.black]
line-length = 100
target-version = ["py39"]

[tool.isort]
profile = "black"
line_length = 100
